# [perf] skip `optimize-attr-pats` for child queries