@stopachka 2024-12-11
Going to do some more research before landing this, as the backtest isn't too promising for the rest of the queries